refactor: 도메인 내부 controller 구조로 라우터 재배치
- interfaces 계층에 있던 학생/교사 라우터를 각 도메인 controller로 이동
- classroom controller DTO를 도메인 내부 dto 패키지로 이동
- main 라우터 import와 테스트 import를 새 구조에 맞게 갱신
- persona 재설계 문서를 도메인 중심 패키징 방향으로 수정
