+++ app/domains/classroom/controller/__init__.py
@@ -1,0 +1,1 @@
+"""Classroom HTTP controllers."""
