+++ app/domains/classroom/controller/dto/__init__.py
@@ -1,0 +1,1 @@
+"""Classroom controller DTOs."""
