+++ app/domains/learning/controller/__init__.py
@@ -1,0 +1,1 @@
+"""Learning HTTP controllers."""
