+++ app/domains/learning/controller/content_catalog_router.py
