--- app/interfaces/content/__init__.py
@@ -1,1 +1,0 @@
-"""Content HTTP interfaces."""
