--- app/interfaces/student/__init__.py
@@ -1,1 +1,0 @@
-
