+++ docs/api-spec.md
@@ -4,7 +4,7 @@
 > 대상: 프론트엔드 / 프론트 에이전트의 리팩토링 작업
 > 관련 문서: [프로젝트 현황](./project-status.md) · [Agent 설계](./ai-agent-design.md)
 
-이 문서는 백엔드가 노출하는 모든 HTTP 엔드포인트의 **요청/응답 스키마**, **인증 요구사항**, **에러 케이스**를 정리한다. 코드 상의 권위 있는 HTTP 정의는 `app/interfaces/**/` 라우터이고, 도메인 스키마는 `app/domains/**/schemas.py`에 있다.
+이 문서는 백엔드가 노출하는 모든 HTTP 엔드포인트의 **요청/응답 스키마**, **인증 요구사항**, **에러 케이스**를 정리한다. 코드 상의 권위 있는 HTTP 정의는 각 도메인의 `controller/` 라우터이고, 도메인 스키마는 `app/domains/**/schemas.py`에 있다.
 
 ---
