+++ tests/test_content_tree_router.py
@@ -3,7 +3,7 @@ from datetime import datetime, timezone
 
 from app.domains.auth.models import User
 from app.domains.learning.content.models import Lesson, Unit
-from app.interfaces.content.catalog_router import get_content_lesson, get_content_units
+from app.domains.learning.controller.content_catalog_router import get_content_lesson, get_content_units
 
 
 def _user() -> User:
