+++ tests/test_domain_structure.py
@@ -14,6 +14,11 @@ def test_domain_packages_are_importable():
         "app.domains.auth.router",
         "app.domains.auth.schemas",
         "app.domains.auth.service",
+        "app.domains.classroom.controller",
+        "app.domains.classroom.controller.dto",
+        "app.domains.classroom.controller.dto.response",
+        "app.domains.classroom.controller.teacher_classroom_router",
+        "app.domains.classroom.controller.teacher_student_view_router",
         "app.domains.classroom.dependencies",
         "app.domains.classroom.models",
         "app.domains.classroom.repositories.base",
@@ -24,6 +29,12 @@ def test_domain_packages_are_importable():
         "app.domains.chat.schemas",
         "app.domains.chat.service",
         "app.domains.learning",
+        "app.domains.learning.controller",
+        "app.domains.learning.controller.content_catalog_router",
+        "app.domains.learning.controller.student_learning_router",
+        "app.domains.learning.controller.student_progress_router",
+        "app.domains.learning.controller.student_records_router",
+        "app.domains.learning.controller.student_stage3_router",
         "app.domains.learning.content",
         "app.domains.learning.content.models",
         "app.domains.learning.content.repositories.base",
@@ -47,18 +58,6 @@ def test_domain_packages_are_importable():
         "app.domains.rag.retriever",
         "app.domains.rag.schemas",
         "app.domains.rag.service",
-        "app.interfaces",
-        "app.interfaces.content",
-        "app.interfaces.content.catalog_router",
-        "app.interfaces.student",
-        "app.interfaces.student.learning_router",
-        "app.interfaces.student.progress_router",
-        "app.interfaces.student.records_router",
-        "app.interfaces.student.stage3_router",
-        "app.interfaces.teacher",
-        "app.interfaces.teacher.classroom_router",
-        "app.interfaces.teacher.schemas",
-        "app.interfaces.teacher.student_view_router",
     ]
 
     for module_name in modules:
